requirements.txt

Django==4.2.5
django-autoslug
django-restframework==0.0.1
djangorestframework==3.14.0
Pillow==9.5.0
